core code updates
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@5385 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/persist/TableSpaceManager.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2001-2011, The HSQL Development Group
+/* Copyright (c) 2001-2014, The HSQL Development Group
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -52,7 +52,7 @@ public interface TableSpaceManager {
     /**
      * Allocates a block
      */
-    long getFilePosition(long rowSize, boolean asBlocks);
+    long getFilePosition(int rowSize, boolean asBlocks);
 
     boolean hasFileRoom(long blockSize);
